fix(ts): Add full test coverage for TypeScript hook types & fix missing overloads (#895)
* fix: add apollo-composable type overloads + tests

There were a lot of TypeScript edge cases, where calling functions with
certain argument options would result in `any` being expressed as a
type, or where the variables/options were not strictly requiring the
desired inputs.

This adds TypeScript function overloads for all hook edge cases so that
all types are correct. This does not change any behavior.

This also adds almost complete coverage for types, excepting the cases
where `strict` is required, or where a failure is expected, which
TypeScript does not currently support. See:
https://github.com/microsoft/TypeScript/issues/29394

* chore: run type tests on prepublish

* chore: code style